Analyze Implement tp-analyze drop tp-recommend
Let the tp-analyze command be used for all data analysis. Use NLTK [1]
for much of the back-end.

Drop tp-recommend because of time constraints, and because tp-analyze
alone is good enough to produce useful results in a timely manner.

There are some issues with the datasets used by the application. Read
the README.

[1] http://www.nltk.org/