Optimize tweet munging code
The order of non-header rows in ``ExtractedTweets.csv`` has no impact on
program correctness. Make the code which munges this file write
corrected rows in an arbitrary order, thus slightly improving
performance.

There are other far greater performance enhancements to be had, but this
is a one-line easy-to-understand change, so why not?