Create btrfs snapshots hourly, not daily
Taking snapshots every hour instead of every day means that a user
should lose up to an hour of work when things go wrong, instead of a
day's worth of work. In addition, this should have little cost, as
snapshots themselves are supposed to be light-weight. (If implementation
is poor, this could turn out to be false.)