Replace aur role with custom-repo, pacman-conf
The aur role created a custom local repository, compiled packages to it,
and made pacman get packages from it. It's a nice all-in-one solution.
But I'm switching to a client-server model.

The new custom-repo role creates a custom repository and compiles
packages to it. It doesn't *serve* the custom repository, but other
roles (like nginx) can easily do this. It's targeted at servers.

The pacman-conf role configures /etc/pacman.conf and
/etc/pacman.d/*.conf. It's targeted at clients. The update-mirrorlist
role (which manages /etc/pacman.d/mirrorlist) may be merged into it.

The aur role is still around, but it just deletes what the aur role used
to do.