Rewrite custom-repo to handle multiple repos
Currently, custom-repo can only create and manage a single repository.
It can be used to create and manage multiple repositories, but only by
calling it multiple times with different variables each time. This
structure harms the goal of separation of concerns: it's desirable to
place all variables in locations such as `group_vars/` and `host_vars/`,
but in that case, a role can't be called multiple times with differing
variables.

Restructure custom-repo so that one can create and manage multiple
repositories. This was done primarily by making the role accept
different variables, and then looping over those variables.