Delete convoluted system monitoring code
System monitoring is important, and it should be in place. However, the
current implementation of system monitoring is overly complicated and is
itself error-prone. Drop all system monitoring code, so that effort can
be put in to improving other existing code, and in to adopting a more
sustainable system monitoring technique.